Issue #259 Hide GAE API from webapp classpath
Handles #259 bu creating AppengineApiConfiguration to manage the classpath that is exposed and/or protected from
the webapplication.    If the GAE API is not available in the webapp, then a warning is logged and the API is
added to the classpath.

Currently javax.activation and javax.mail are treated as shared classes and the containers version is made available.
If these are not listed, then even if they are in the added API jar, errors occur, which needs to be investigated.

Currently a specific list of org.apache packages are hidden.   A better approach may be to hide all org.apache classes
and selectively reveals those needed for JSP, JSTL and EL support?